Time: 3 ms (36.58%), Space: 5.9 MB (74.34%) - LeetHub

+++ 62-unique-paths/62-unique-paths.cpp
@@ -16,7 +16,7 @@ public:
         
         int N = n + m - 2;
         int r = m - 1;
-        double res = 1;
+        long int res = 1;
         for (int i = 1; i <= r; i++)
             res = res * (N-r+i)/i;
